Subwindow sizes can be changed.
The GridFlex.m added for windows that have 2 halves and a line for
changed the size of the subwindows. This is a in the original
application. Due to only the uifigure object have the window button
Down, Up, and Motions Callbacks, other files needed to be changed so
these callbacks are added the the parent figure.